Refactor `RedisTest`
- Rename to be symmertic to service name.
- Remove unneeded getters in test class.
- Remove unneeded Mocks.
- Disable success test to avoid crossing free Twilio limit.
