+++ src/test/java/com/ittovative/otpservice/OtpServiceApplicationTests.java
@@ -1,18 +1,15 @@
 package com.ittovative.otpservice;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-/**
- * The type Otp service application tests.
- */
 @SpringBootTest
+@DisplayName("Main Application")
 class OtpServiceApplicationTests {
 
-    /**
-     * Context loads.
-     */
     @Test
+    @DisplayName("Load application context")
     void contextLoads() {
     }
 }
